+++ RELEASES.md
@@ -3,6 +3,7 @@
 ### Bug fixes
 
 -   Temporary, but effective, fix for the bars overflowing from their container when there's a fullscreen window [[#127](https://github.com/AstraExt/astra-monitor/issues/127)]
+-   Addressed a layout issue in horizontal sensors header where sudden changes in one of the sensors values length could disrupt the overall arrangement [[#111](https://github.com/AstraExt/astra-monitor/issues/111)]
 -   Resolved an issue where `iotop` was not accurately identifying process names when both SWAPIN and IO% information were present
 
 # Astra Monitor 28 - September 27 2024